backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
opensearch-py==2.4.2
pandas>=2.1.3
python-dotenv>=1.0.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
numpy>=1.24.3
requests==2.31.0
scikit-learn>=0.24.0
matplotlib>=3.5.0
seaborn>=0.11.0
jupyter>=1.0.0
pymongo>=4.0.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
redis==5.0.1
kafka-python==2.0.2
python-json-logger==2.0.7
boto3>=1.26.0
requests-aws4auth>=1.2.0
slowapi==0.1.9
